server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.6.0",
    "cryptography>=46.0.3",
    "django>=5.2.8",
    "django-filter>=25.2",
    "django-ipware>=7.0.1",
    "django-otp>=1.6.3",
    "django-redis>=6.0.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.2.1",
    "ua-parser>=1.0.1",
    "user-agents>=2.2.0",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "pre-commit",
]
